+++ protoc_plugin/lib/message_generator.dart
@@ -449,6 +449,7 @@ class MessageGenerator extends ProtobufContainer {
 
     _emitDeprecatedIf(field.isDeprecated, out);
     _emitOverrideIf(field.overridesGetter, out);
+    _emitIndexAnnotation(field.number, out);
     final getterExpr = _getterExpression(fieldTypeString, field.index,
         defaultExpr, field.isRepeated, field.isMapField);
     out.printlnAnnotated(
@@ -476,6 +477,7 @@ class MessageGenerator extends ProtobufContainer {
       var fastSetter = field.baseType.setter;
       _emitDeprecatedIf(field.isDeprecated, out);
       _emitOverrideIf(field.overridesSetter, out);
+      _emitIndexAnnotation(field.number, out);
       if (fastSetter != null) {
         out.printlnAnnotated(
             'set ${names.fieldName}'
@@ -503,6 +505,7 @@ class MessageGenerator extends ProtobufContainer {
       }
       _emitDeprecatedIf(field.isDeprecated, out);
       _emitOverrideIf(field.overridesHasMethod, out);
+      _emitIndexAnnotation(field.number, out);
       out.printlnAnnotated(
           '$_coreImportPrefix.bool ${names.hasMethodName}() =>'
           ' \$_has(${field.index});',
@@ -514,6 +517,7 @@ class MessageGenerator extends ProtobufContainer {
           ]);
       _emitDeprecatedIf(field.isDeprecated, out);
       _emitOverrideIf(field.overridesClearMethod, out);
+      _emitIndexAnnotation(field.number, out);
       out.printlnAnnotated(
           'void ${names.clearMethodName}() =>'
           ' clearField(${field.number});',
@@ -554,6 +558,10 @@ class MessageGenerator extends ProtobufContainer {
     if (condition) {
       out.println('@$_coreImportPrefix.override');
     }
+  }
+
+  void _emitIndexAnnotation(int index, IndentingWriter out) {
+    out.println('@$_protobufImportPrefix.TagNumber($index)');
   }
 
   void generateEnums(IndentingWriter out) {