+++ protoc_plugin/lib/src/descriptor.pb.dart
@@ -16,7 +16,8 @@ export 'descriptor.pbenum.dart';
 
 class FileDescriptorSet extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('FileDescriptorSet',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..pc<FileDescriptorProto>(1, 'file', $pb.PbFieldType.PM,
         subBuilder: FileDescriptorProto.create);
 
@@ -41,12 +42,14 @@ class FileDescriptorSet extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static FileDescriptorSet _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.List<FileDescriptorProto> get file => $_getList(0);
 }
 
 class FileDescriptorProto extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('FileDescriptorProto',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOS(1, 'name')
     ..aOS(2, 'package')
     ..pPS(3, 'dependency')
@@ -88,65 +91,93 @@ class FileDescriptorProto extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static FileDescriptorProto _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get name => $_getS(0, '');
+  @$pb.TagNumber(1)
   set name($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasName() => $_has(0);
+  @$pb.TagNumber(1)
   void clearName() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.String get package => $_getS(1, '');
+  @$pb.TagNumber(2)
   set package($core.String v) {
     $_setString(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasPackage() => $_has(1);
+  @$pb.TagNumber(2)
   void clearPackage() => clearField(2);
 
+  @$pb.TagNumber(3)
   $core.List<$core.String> get dependency => $_getList(2);
 
+  @$pb.TagNumber(4)
   $core.List<DescriptorProto> get messageType => $_getList(3);
 
+  @$pb.TagNumber(5)
   $core.List<EnumDescriptorProto> get enumType => $_getList(4);
 
+  @$pb.TagNumber(6)
   $core.List<ServiceDescriptorProto> get service => $_getList(5);
 
+  @$pb.TagNumber(7)
   $core.List<FieldDescriptorProto> get extension => $_getList(6);
 
+  @$pb.TagNumber(8)
   FileOptions get options => $_getN(7);
+  @$pb.TagNumber(8)
   set options(FileOptions v) {
     setField(8, v);
   }
 
+  @$pb.TagNumber(8)
   $core.bool hasOptions() => $_has(7);
+  @$pb.TagNumber(8)
   void clearOptions() => clearField(8);
 
+  @$pb.TagNumber(9)
   SourceCodeInfo get sourceCodeInfo => $_getN(8);
+  @$pb.TagNumber(9)
   set sourceCodeInfo(SourceCodeInfo v) {
     setField(9, v);
   }
 
+  @$pb.TagNumber(9)
   $core.bool hasSourceCodeInfo() => $_has(8);
+  @$pb.TagNumber(9)
   void clearSourceCodeInfo() => clearField(9);
 
+  @$pb.TagNumber(10)
   $core.List<$core.int> get publicDependency => $_getList(9);
 
+  @$pb.TagNumber(11)
   $core.List<$core.int> get weakDependency => $_getList(10);
 
+  @$pb.TagNumber(12)
   $core.String get syntax => $_getS(11, '');
+  @$pb.TagNumber(12)
   set syntax($core.String v) {
     $_setString(11, v);
   }
 
+  @$pb.TagNumber(12)
   $core.bool hasSyntax() => $_has(11);
+  @$pb.TagNumber(12)
   void clearSyntax() => clearField(12);
 }
 
 class DescriptorProto_ExtensionRange extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo(
       'DescriptorProto.ExtensionRange',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..a<$core.int>(1, 'start', $pb.PbFieldType.O3)
     ..a<$core.int>(2, 'end', $pb.PbFieldType.O3)
     ..hasRequiredFields = false;
@@ -176,27 +207,36 @@ class DescriptorProto_ExtensionRange extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static DescriptorProto_ExtensionRange _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.int get start => $_get(0, 0);
+  @$pb.TagNumber(1)
   set start($core.int v) {
     $_setSignedInt32(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasStart() => $_has(0);
+  @$pb.TagNumber(1)
   void clearStart() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.int get end => $_get(1, 0);
+  @$pb.TagNumber(2)
   set end($core.int v) {
     $_setSignedInt32(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasEnd() => $_has(1);
+  @$pb.TagNumber(2)
   void clearEnd() => clearField(2);
 }
 
 class DescriptorProto_ReservedRange extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo(
       'DescriptorProto.ReservedRange',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..a<$core.int>(1, 'start', $pb.PbFieldType.O3)
     ..a<$core.int>(2, 'end', $pb.PbFieldType.O3)
     ..hasRequiredFields = false;
@@ -226,26 +266,35 @@ class DescriptorProto_ReservedRange extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static DescriptorProto_ReservedRange _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.int get start => $_get(0, 0);
+  @$pb.TagNumber(1)
   set start($core.int v) {
     $_setSignedInt32(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasStart() => $_has(0);
+  @$pb.TagNumber(1)
   void clearStart() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.int get end => $_get(1, 0);
+  @$pb.TagNumber(2)
   set end($core.int v) {
     $_setSignedInt32(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasEnd() => $_has(1);
+  @$pb.TagNumber(2)
   void clearEnd() => clearField(2);
 }
 
 class DescriptorProto extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('DescriptorProto',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOS(1, 'name')
     ..pc<FieldDescriptorProto>(2, 'field', $pb.PbFieldType.PM,
         subBuilder: FieldDescriptorProto.create)
@@ -288,42 +337,59 @@ class DescriptorProto extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static DescriptorProto _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get name => $_getS(0, '');
+  @$pb.TagNumber(1)
   set name($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasName() => $_has(0);
+  @$pb.TagNumber(1)
   void clearName() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.List<FieldDescriptorProto> get field => $_getList(1);
 
+  @$pb.TagNumber(3)
   $core.List<DescriptorProto> get nestedType => $_getList(2);
 
+  @$pb.TagNumber(4)
   $core.List<EnumDescriptorProto> get enumType => $_getList(3);
 
+  @$pb.TagNumber(5)
   $core.List<DescriptorProto_ExtensionRange> get extensionRange => $_getList(4);
 
+  @$pb.TagNumber(6)
   $core.List<FieldDescriptorProto> get extension => $_getList(5);
 
+  @$pb.TagNumber(7)
   MessageOptions get options => $_getN(6);
+  @$pb.TagNumber(7)
   set options(MessageOptions v) {
     setField(7, v);
   }
 
+  @$pb.TagNumber(7)
   $core.bool hasOptions() => $_has(6);
+  @$pb.TagNumber(7)
   void clearOptions() => clearField(7);
 
+  @$pb.TagNumber(8)
   $core.List<OneofDescriptorProto> get oneofDecl => $_getList(7);
 
+  @$pb.TagNumber(9)
   $core.List<DescriptorProto_ReservedRange> get reservedRange => $_getList(8);
 
+  @$pb.TagNumber(10)
   $core.List<$core.String> get reservedName => $_getList(9);
 }
 
 class FieldDescriptorProto extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('FieldDescriptorProto',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOS(1, 'name')
     ..aOS(2, 'extendee')
     ..a<$core.int>(3, 'number', $pb.PbFieldType.O3)
@@ -365,90 +431,131 @@ class FieldDescriptorProto extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static FieldDescriptorProto _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get name => $_getS(0, '');
+  @$pb.TagNumber(1)
   set name($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasName() => $_has(0);
+  @$pb.TagNumber(1)
   void clearName() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.String get extendee => $_getS(1, '');
+  @$pb.TagNumber(2)
   set extendee($core.String v) {
     $_setString(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasExtendee() => $_has(1);
+  @$pb.TagNumber(2)
   void clearExtendee() => clearField(2);
 
+  @$pb.TagNumber(3)
   $core.int get number => $_get(2, 0);
+  @$pb.TagNumber(3)
   set number($core.int v) {
     $_setSignedInt32(2, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasNumber() => $_has(2);
+  @$pb.TagNumber(3)
   void clearNumber() => clearField(3);
 
+  @$pb.TagNumber(4)
   FieldDescriptorProto_Label get label => $_getN(3);
+  @$pb.TagNumber(4)
   set label(FieldDescriptorProto_Label v) {
     setField(4, v);
   }
 
+  @$pb.TagNumber(4)
   $core.bool hasLabel() => $_has(3);
+  @$pb.TagNumber(4)
   void clearLabel() => clearField(4);
 
+  @$pb.TagNumber(5)
   FieldDescriptorProto_Type get type => $_getN(4);
+  @$pb.TagNumber(5)
   set type(FieldDescriptorProto_Type v) {
     setField(5, v);
   }
 
+  @$pb.TagNumber(5)
   $core.bool hasType() => $_has(4);
+  @$pb.TagNumber(5)
   void clearType() => clearField(5);
 
+  @$pb.TagNumber(6)
   $core.String get typeName => $_getS(5, '');
+  @$pb.TagNumber(6)
   set typeName($core.String v) {
     $_setString(5, v);
   }
 
+  @$pb.TagNumber(6)
   $core.bool hasTypeName() => $_has(5);
+  @$pb.TagNumber(6)
   void clearTypeName() => clearField(6);
 
+  @$pb.TagNumber(7)
   $core.String get defaultValue => $_getS(6, '');
+  @$pb.TagNumber(7)
   set defaultValue($core.String v) {
     $_setString(6, v);
   }
 
+  @$pb.TagNumber(7)
   $core.bool hasDefaultValue() => $_has(6);
+  @$pb.TagNumber(7)
   void clearDefaultValue() => clearField(7);
 
+  @$pb.TagNumber(8)
   FieldOptions get options => $_getN(7);
+  @$pb.TagNumber(8)
   set options(FieldOptions v) {
     setField(8, v);
   }
 
+  @$pb.TagNumber(8)
   $core.bool hasOptions() => $_has(7);
+  @$pb.TagNumber(8)
   void clearOptions() => clearField(8);
 
+  @$pb.TagNumber(9)
   $core.int get oneofIndex => $_get(8, 0);
+  @$pb.TagNumber(9)
   set oneofIndex($core.int v) {
     $_setSignedInt32(8, v);
   }
 
+  @$pb.TagNumber(9)
   $core.bool hasOneofIndex() => $_has(8);
+  @$pb.TagNumber(9)
   void clearOneofIndex() => clearField(9);
 
+  @$pb.TagNumber(10)
   $core.String get jsonName => $_getS(9, '');
+  @$pb.TagNumber(10)
   set jsonName($core.String v) {
     $_setString(9, v);
   }
 
+  @$pb.TagNumber(10)
   $core.bool hasJsonName() => $_has(9);
+  @$pb.TagNumber(10)
   void clearJsonName() => clearField(10);
 }
 
 class OneofDescriptorProto extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('OneofDescriptorProto',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOS(1, 'name')
     ..a<OneofOptions>(2, 'options', $pb.PbFieldType.OM,
         defaultOrMaker: OneofOptions.getDefault,
@@ -476,26 +583,35 @@ class OneofDescriptorProto extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static OneofDescriptorProto _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get name => $_getS(0, '');
+  @$pb.TagNumber(1)
   set name($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasName() => $_has(0);
+  @$pb.TagNumber(1)
   void clearName() => clearField(1);
 
+  @$pb.TagNumber(2)
   OneofOptions get options => $_getN(1);
+  @$pb.TagNumber(2)
   set options(OneofOptions v) {
     setField(2, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasOptions() => $_has(1);
+  @$pb.TagNumber(2)
   void clearOptions() => clearField(2);
 }
 
 class EnumDescriptorProto extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('EnumDescriptorProto',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOS(1, 'name')
     ..pc<EnumValueDescriptorProto>(2, 'value', $pb.PbFieldType.PM,
         subBuilder: EnumValueDescriptorProto.create)
@@ -523,28 +639,38 @@ class EnumDescriptorProto extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static EnumDescriptorProto _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get name => $_getS(0, '');
+  @$pb.TagNumber(1)
   set name($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasName() => $_has(0);
+  @$pb.TagNumber(1)
   void clearName() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.List<EnumValueDescriptorProto> get value => $_getList(1);
 
+  @$pb.TagNumber(3)
   EnumOptions get options => $_getN(2);
+  @$pb.TagNumber(3)
   set options(EnumOptions v) {
     setField(3, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasOptions() => $_has(2);
+  @$pb.TagNumber(3)
   void clearOptions() => clearField(3);
 }
 
 class EnumValueDescriptorProto extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('EnumValueDescriptorProto',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOS(1, 'name')
     ..a<$core.int>(2, 'number', $pb.PbFieldType.O3)
     ..a<EnumValueOptions>(3, 'options', $pb.PbFieldType.OM,
@@ -574,34 +700,47 @@ class EnumValueDescriptorProto extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static EnumValueDescriptorProto _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get name => $_getS(0, '');
+  @$pb.TagNumber(1)
   set name($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasName() => $_has(0);
+  @$pb.TagNumber(1)
   void clearName() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.int get number => $_get(1, 0);
+  @$pb.TagNumber(2)
   set number($core.int v) {
     $_setSignedInt32(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasNumber() => $_has(1);
+  @$pb.TagNumber(2)
   void clearNumber() => clearField(2);
 
+  @$pb.TagNumber(3)
   EnumValueOptions get options => $_getN(2);
+  @$pb.TagNumber(3)
   set options(EnumValueOptions v) {
     setField(3, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasOptions() => $_has(2);
+  @$pb.TagNumber(3)
   void clearOptions() => clearField(3);
 }
 
 class ServiceDescriptorProto extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('ServiceDescriptorProto',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOS(1, 'name')
     ..pc<MethodDescriptorProto>(2, 'method', $pb.PbFieldType.PM,
         subBuilder: MethodDescriptorProto.create)
@@ -632,28 +771,38 @@ class ServiceDescriptorProto extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static ServiceDescriptorProto _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get name => $_getS(0, '');
+  @$pb.TagNumber(1)
   set name($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasName() => $_has(0);
+  @$pb.TagNumber(1)
   void clearName() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.List<MethodDescriptorProto> get method => $_getList(1);
 
+  @$pb.TagNumber(3)
   ServiceOptions get options => $_getN(2);
+  @$pb.TagNumber(3)
   set options(ServiceOptions v) {
     setField(3, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasOptions() => $_has(2);
+  @$pb.TagNumber(3)
   void clearOptions() => clearField(3);
 }
 
 class MethodDescriptorProto extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('MethodDescriptorProto',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOS(1, 'name')
     ..aOS(2, 'inputType')
     ..aOS(3, 'outputType')
@@ -686,58 +835,83 @@ class MethodDescriptorProto extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static MethodDescriptorProto _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get name => $_getS(0, '');
+  @$pb.TagNumber(1)
   set name($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasName() => $_has(0);
+  @$pb.TagNumber(1)
   void clearName() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.String get inputType => $_getS(1, '');
+  @$pb.TagNumber(2)
   set inputType($core.String v) {
     $_setString(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasInputType() => $_has(1);
+  @$pb.TagNumber(2)
   void clearInputType() => clearField(2);
 
+  @$pb.TagNumber(3)
   $core.String get outputType => $_getS(2, '');
+  @$pb.TagNumber(3)
   set outputType($core.String v) {
     $_setString(2, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasOutputType() => $_has(2);
+  @$pb.TagNumber(3)
   void clearOutputType() => clearField(3);
 
+  @$pb.TagNumber(4)
   MethodOptions get options => $_getN(3);
+  @$pb.TagNumber(4)
   set options(MethodOptions v) {
     setField(4, v);
   }
 
+  @$pb.TagNumber(4)
   $core.bool hasOptions() => $_has(3);
+  @$pb.TagNumber(4)
   void clearOptions() => clearField(4);
 
+  @$pb.TagNumber(5)
   $core.bool get clientStreaming => $_get(4, false);
+  @$pb.TagNumber(5)
   set clientStreaming($core.bool v) {
     $_setBool(4, v);
   }
 
+  @$pb.TagNumber(5)
   $core.bool hasClientStreaming() => $_has(4);
+  @$pb.TagNumber(5)
   void clearClientStreaming() => clearField(5);
 
+  @$pb.TagNumber(6)
   $core.bool get serverStreaming => $_get(5, false);
+  @$pb.TagNumber(6)
   set serverStreaming($core.bool v) {
     $_setBool(5, v);
   }
 
+  @$pb.TagNumber(6)
   $core.bool hasServerStreaming() => $_has(5);
+  @$pb.TagNumber(6)
   void clearServerStreaming() => clearField(6);
 }
 
 class FileOptions extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('FileOptions',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOS(1, 'javaPackage')
     ..aOS(8, 'javaOuterClassname')
     ..e<FileOptions_OptimizeMode>(9, 'optimizeFor', $pb.PbFieldType.OE,
@@ -781,152 +955,222 @@ class FileOptions extends $pb.GeneratedMessage {
   static FileOptions getDefault() => _defaultInstance ??= create()..freeze();
   static FileOptions _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get javaPackage => $_getS(0, '');
+  @$pb.TagNumber(1)
   set javaPackage($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasJavaPackage() => $_has(0);
+  @$pb.TagNumber(1)
   void clearJavaPackage() => clearField(1);
 
+  @$pb.TagNumber(8)
   $core.String get javaOuterClassname => $_getS(1, '');
+  @$pb.TagNumber(8)
   set javaOuterClassname($core.String v) {
     $_setString(1, v);
   }
 
+  @$pb.TagNumber(8)
   $core.bool hasJavaOuterClassname() => $_has(1);
+  @$pb.TagNumber(8)
   void clearJavaOuterClassname() => clearField(8);
 
+  @$pb.TagNumber(9)
   FileOptions_OptimizeMode get optimizeFor => $_getN(2);
+  @$pb.TagNumber(9)
   set optimizeFor(FileOptions_OptimizeMode v) {
     setField(9, v);
   }
 
+  @$pb.TagNumber(9)
   $core.bool hasOptimizeFor() => $_has(2);
+  @$pb.TagNumber(9)
   void clearOptimizeFor() => clearField(9);
 
+  @$pb.TagNumber(10)
   $core.bool get javaMultipleFiles => $_get(3, false);
+  @$pb.TagNumber(10)
   set javaMultipleFiles($core.bool v) {
     $_setBool(3, v);
   }
 
+  @$pb.TagNumber(10)
   $core.bool hasJavaMultipleFiles() => $_has(3);
+  @$pb.TagNumber(10)
   void clearJavaMultipleFiles() => clearField(10);
 
+  @$pb.TagNumber(11)
   $core.String get goPackage => $_getS(4, '');
+  @$pb.TagNumber(11)
   set goPackage($core.String v) {
     $_setString(4, v);
   }
 
+  @$pb.TagNumber(11)
   $core.bool hasGoPackage() => $_has(4);
+  @$pb.TagNumber(11)
   void clearGoPackage() => clearField(11);
 
+  @$pb.TagNumber(16)
   $core.bool get ccGenericServices => $_get(5, false);
+  @$pb.TagNumber(16)
   set ccGenericServices($core.bool v) {
     $_setBool(5, v);
   }
 
+  @$pb.TagNumber(16)
   $core.bool hasCcGenericServices() => $_has(5);
+  @$pb.TagNumber(16)
   void clearCcGenericServices() => clearField(16);
 
+  @$pb.TagNumber(17)
   $core.bool get javaGenericServices => $_get(6, false);
+  @$pb.TagNumber(17)
   set javaGenericServices($core.bool v) {
     $_setBool(6, v);
   }
 
+  @$pb.TagNumber(17)
   $core.bool hasJavaGenericServices() => $_has(6);
+  @$pb.TagNumber(17)
   void clearJavaGenericServices() => clearField(17);
 
+  @$pb.TagNumber(18)
   $core.bool get pyGenericServices => $_get(7, false);
+  @$pb.TagNumber(18)
   set pyGenericServices($core.bool v) {
     $_setBool(7, v);
   }
 
+  @$pb.TagNumber(18)
   $core.bool hasPyGenericServices() => $_has(7);
+  @$pb.TagNumber(18)
   void clearPyGenericServices() => clearField(18);
 
   @$core.Deprecated('This field is deprecated.')
+  @$pb.TagNumber(20)
   $core.bool get javaGenerateEqualsAndHash => $_get(8, false);
   @$core.Deprecated('This field is deprecated.')
+  @$pb.TagNumber(20)
   set javaGenerateEqualsAndHash($core.bool v) {
     $_setBool(8, v);
   }
 
   @$core.Deprecated('This field is deprecated.')
+  @$pb.TagNumber(20)
   $core.bool hasJavaGenerateEqualsAndHash() => $_has(8);
   @$core.Deprecated('This field is deprecated.')
+  @$pb.TagNumber(20)
   void clearJavaGenerateEqualsAndHash() => clearField(20);
 
+  @$pb.TagNumber(23)
   $core.bool get deprecated => $_get(9, false);
+  @$pb.TagNumber(23)
   set deprecated($core.bool v) {
     $_setBool(9, v);
   }
 
+  @$pb.TagNumber(23)
   $core.bool hasDeprecated() => $_has(9);
+  @$pb.TagNumber(23)
   void clearDeprecated() => clearField(23);
 
+  @$pb.TagNumber(27)
   $core.bool get javaStringCheckUtf8 => $_get(10, false);
+  @$pb.TagNumber(27)
   set javaStringCheckUtf8($core.bool v) {
     $_setBool(10, v);
   }
 
+  @$pb.TagNumber(27)
   $core.bool hasJavaStringCheckUtf8() => $_has(10);
+  @$pb.TagNumber(27)
   void clearJavaStringCheckUtf8() => clearField(27);
 
+  @$pb.TagNumber(31)
   $core.bool get ccEnableArenas => $_get(11, false);
+  @$pb.TagNumber(31)
   set ccEnableArenas($core.bool v) {
     $_setBool(11, v);
   }
 
+  @$pb.TagNumber(31)
   $core.bool hasCcEnableArenas() => $_has(11);
+  @$pb.TagNumber(31)
   void clearCcEnableArenas() => clearField(31);
 
+  @$pb.TagNumber(36)
   $core.String get objcClassPrefix => $_getS(12, '');
+  @$pb.TagNumber(36)
   set objcClassPrefix($core.String v) {
     $_setString(12, v);
   }
 
+  @$pb.TagNumber(36)
   $core.bool hasObjcClassPrefix() => $_has(12);
+  @$pb.TagNumber(36)
   void clearObjcClassPrefix() => clearField(36);
 
+  @$pb.TagNumber(37)
   $core.String get csharpNamespace => $_getS(13, '');
+  @$pb.TagNumber(37)
   set csharpNamespace($core.String v) {
     $_setString(13, v);
   }
 
+  @$pb.TagNumber(37)
   $core.bool hasCsharpNamespace() => $_has(13);
+  @$pb.TagNumber(37)
   void clearCsharpNamespace() => clearField(37);
 
+  @$pb.TagNumber(39)
   $core.String get swiftPrefix => $_getS(14, '');
+  @$pb.TagNumber(39)
   set swiftPrefix($core.String v) {
     $_setString(14, v);
   }
 
+  @$pb.TagNumber(39)
   $core.bool hasSwiftPrefix() => $_has(14);
+  @$pb.TagNumber(39)
   void clearSwiftPrefix() => clearField(39);
 
+  @$pb.TagNumber(40)
   $core.String get phpClassPrefix => $_getS(15, '');
+  @$pb.TagNumber(40)
   set phpClassPrefix($core.String v) {
     $_setString(15, v);
   }
 
+  @$pb.TagNumber(40)
   $core.bool hasPhpClassPrefix() => $_has(15);
+  @$pb.TagNumber(40)
   void clearPhpClassPrefix() => clearField(40);
 
+  @$pb.TagNumber(41)
   $core.String get phpNamespace => $_getS(16, '');
+  @$pb.TagNumber(41)
   set phpNamespace($core.String v) {
     $_setString(16, v);
   }
 
+  @$pb.TagNumber(41)
   $core.bool hasPhpNamespace() => $_has(16);
+  @$pb.TagNumber(41)
   void clearPhpNamespace() => clearField(41);
 
+  @$pb.TagNumber(999)
   $core.List<UninterpretedOption> get uninterpretedOption => $_getList(17);
 }
 
 class MessageOptions extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('MessageOptions',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOB(1, 'messageSetWireFormat')
     ..aOB(2, 'noStandardDescriptorAccessor')
     ..aOB(3, 'deprecated')
@@ -955,44 +1199,62 @@ class MessageOptions extends $pb.GeneratedMessage {
   static MessageOptions getDefault() => _defaultInstance ??= create()..freeze();
   static MessageOptions _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.bool get messageSetWireFormat => $_get(0, false);
+  @$pb.TagNumber(1)
   set messageSetWireFormat($core.bool v) {
     $_setBool(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasMessageSetWireFormat() => $_has(0);
+  @$pb.TagNumber(1)
   void clearMessageSetWireFormat() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.bool get noStandardDescriptorAccessor => $_get(1, false);
+  @$pb.TagNumber(2)
   set noStandardDescriptorAccessor($core.bool v) {
     $_setBool(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasNoStandardDescriptorAccessor() => $_has(1);
+  @$pb.TagNumber(2)
   void clearNoStandardDescriptorAccessor() => clearField(2);
 
+  @$pb.TagNumber(3)
   $core.bool get deprecated => $_get(2, false);
+  @$pb.TagNumber(3)
   set deprecated($core.bool v) {
     $_setBool(2, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasDeprecated() => $_has(2);
+  @$pb.TagNumber(3)
   void clearDeprecated() => clearField(3);
 
+  @$pb.TagNumber(7)
   $core.bool get mapEntry => $_get(3, false);
+  @$pb.TagNumber(7)
   set mapEntry($core.bool v) {
     $_setBool(3, v);
   }
 
+  @$pb.TagNumber(7)
   $core.bool hasMapEntry() => $_has(3);
+  @$pb.TagNumber(7)
   void clearMapEntry() => clearField(7);
 
+  @$pb.TagNumber(999)
   $core.List<UninterpretedOption> get uninterpretedOption => $_getList(4);
 }
 
 class FieldOptions extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('FieldOptions',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..e<FieldOptions_CType>(1, 'ctype', $pb.PbFieldType.OE,
         defaultOrMaker: FieldOptions_CType.STRING,
         valueOf: FieldOptions_CType.valueOf,
@@ -1029,60 +1291,86 @@ class FieldOptions extends $pb.GeneratedMessage {
   static FieldOptions getDefault() => _defaultInstance ??= create()..freeze();
   static FieldOptions _defaultInstance;
 
+  @$pb.TagNumber(1)
   FieldOptions_CType get ctype => $_getN(0);
+  @$pb.TagNumber(1)
   set ctype(FieldOptions_CType v) {
     setField(1, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasCtype() => $_has(0);
+  @$pb.TagNumber(1)
   void clearCtype() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.bool get packed => $_get(1, false);
+  @$pb.TagNumber(2)
   set packed($core.bool v) {
     $_setBool(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasPacked() => $_has(1);
+  @$pb.TagNumber(2)
   void clearPacked() => clearField(2);
 
+  @$pb.TagNumber(3)
   $core.bool get deprecated => $_get(2, false);
+  @$pb.TagNumber(3)
   set deprecated($core.bool v) {
     $_setBool(2, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasDeprecated() => $_has(2);
+  @$pb.TagNumber(3)
   void clearDeprecated() => clearField(3);
 
+  @$pb.TagNumber(5)
   $core.bool get lazy => $_get(3, false);
+  @$pb.TagNumber(5)
   set lazy($core.bool v) {
     $_setBool(3, v);
   }
 
+  @$pb.TagNumber(5)
   $core.bool hasLazy() => $_has(3);
+  @$pb.TagNumber(5)
   void clearLazy() => clearField(5);
 
+  @$pb.TagNumber(6)
   FieldOptions_JSType get jstype => $_getN(4);
+  @$pb.TagNumber(6)
   set jstype(FieldOptions_JSType v) {
     setField(6, v);
   }
 
+  @$pb.TagNumber(6)
   $core.bool hasJstype() => $_has(4);
+  @$pb.TagNumber(6)
   void clearJstype() => clearField(6);
 
+  @$pb.TagNumber(10)
   $core.bool get weak => $_get(5, false);
+  @$pb.TagNumber(10)
   set weak($core.bool v) {
     $_setBool(5, v);
   }
 
+  @$pb.TagNumber(10)
   $core.bool hasWeak() => $_has(5);
+  @$pb.TagNumber(10)
   void clearWeak() => clearField(10);
 
+  @$pb.TagNumber(999)
   $core.List<UninterpretedOption> get uninterpretedOption => $_getList(6);
 }
 
 class OneofOptions extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('OneofOptions',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..pc<UninterpretedOption>(999, 'uninterpretedOption', $pb.PbFieldType.PM,
         subBuilder: UninterpretedOption.create)
     ..hasExtensions = true;
@@ -1107,12 +1395,14 @@ class OneofOptions extends $pb.GeneratedMessage {
   static OneofOptions getDefault() => _defaultInstance ??= create()..freeze();
   static OneofOptions _defaultInstance;
 
+  @$pb.TagNumber(999)
   $core.List<UninterpretedOption> get uninterpretedOption => $_getList(0);
 }
 
 class EnumOptions extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('EnumOptions',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOB(2, 'allowAlias')
     ..aOB(3, 'deprecated')
     ..pc<UninterpretedOption>(999, 'uninterpretedOption', $pb.PbFieldType.PM,
@@ -1138,28 +1428,38 @@ class EnumOptions extends $pb.GeneratedMessage {
   static EnumOptions getDefault() => _defaultInstance ??= create()..freeze();
   static EnumOptions _defaultInstance;
 
+  @$pb.TagNumber(2)
   $core.bool get allowAlias => $_get(0, false);
+  @$pb.TagNumber(2)
   set allowAlias($core.bool v) {
     $_setBool(0, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasAllowAlias() => $_has(0);
+  @$pb.TagNumber(2)
   void clearAllowAlias() => clearField(2);
 
+  @$pb.TagNumber(3)
   $core.bool get deprecated => $_get(1, false);
+  @$pb.TagNumber(3)
   set deprecated($core.bool v) {
     $_setBool(1, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasDeprecated() => $_has(1);
+  @$pb.TagNumber(3)
   void clearDeprecated() => clearField(3);
 
+  @$pb.TagNumber(999)
   $core.List<UninterpretedOption> get uninterpretedOption => $_getList(2);
 }
 
 class EnumValueOptions extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('EnumValueOptions',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOB(1, 'deprecated')
     ..pc<UninterpretedOption>(999, 'uninterpretedOption', $pb.PbFieldType.PM,
         subBuilder: UninterpretedOption.create)
@@ -1186,20 +1486,26 @@ class EnumValueOptions extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static EnumValueOptions _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.bool get deprecated => $_get(0, false);
+  @$pb.TagNumber(1)
   set deprecated($core.bool v) {
     $_setBool(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasDeprecated() => $_has(0);
+  @$pb.TagNumber(1)
   void clearDeprecated() => clearField(1);
 
+  @$pb.TagNumber(999)
   $core.List<UninterpretedOption> get uninterpretedOption => $_getList(1);
 }
 
 class ServiceOptions extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('ServiceOptions',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOB(33, 'deprecated')
     ..pc<UninterpretedOption>(999, 'uninterpretedOption', $pb.PbFieldType.PM,
         subBuilder: UninterpretedOption.create)
@@ -1225,20 +1531,26 @@ class ServiceOptions extends $pb.GeneratedMessage {
   static ServiceOptions getDefault() => _defaultInstance ??= create()..freeze();
   static ServiceOptions _defaultInstance;
 
+  @$pb.TagNumber(33)
   $core.bool get deprecated => $_get(0, false);
+  @$pb.TagNumber(33)
   set deprecated($core.bool v) {
     $_setBool(0, v);
   }
 
+  @$pb.TagNumber(33)
   $core.bool hasDeprecated() => $_has(0);
+  @$pb.TagNumber(33)
   void clearDeprecated() => clearField(33);
 
+  @$pb.TagNumber(999)
   $core.List<UninterpretedOption> get uninterpretedOption => $_getList(1);
 }
 
 class MethodOptions extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('MethodOptions',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aOB(33, 'deprecated')
     ..e<MethodOptions_IdempotencyLevel>(
         34, 'idempotencyLevel', $pb.PbFieldType.OE,
@@ -1269,29 +1581,39 @@ class MethodOptions extends $pb.GeneratedMessage {
   static MethodOptions getDefault() => _defaultInstance ??= create()..freeze();
   static MethodOptions _defaultInstance;
 
+  @$pb.TagNumber(33)
   $core.bool get deprecated => $_get(0, false);
+  @$pb.TagNumber(33)
   set deprecated($core.bool v) {
     $_setBool(0, v);
   }
 
+  @$pb.TagNumber(33)
   $core.bool hasDeprecated() => $_has(0);
+  @$pb.TagNumber(33)
   void clearDeprecated() => clearField(33);
 
+  @$pb.TagNumber(34)
   MethodOptions_IdempotencyLevel get idempotencyLevel => $_getN(1);
+  @$pb.TagNumber(34)
   set idempotencyLevel(MethodOptions_IdempotencyLevel v) {
     setField(34, v);
   }
 
+  @$pb.TagNumber(34)
   $core.bool hasIdempotencyLevel() => $_has(1);
+  @$pb.TagNumber(34)
   void clearIdempotencyLevel() => clearField(34);
 
+  @$pb.TagNumber(999)
   $core.List<UninterpretedOption> get uninterpretedOption => $_getList(2);
 }
 
 class UninterpretedOption_NamePart extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo(
       'UninterpretedOption.NamePart',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..aQS(1, 'namePart')
     ..a<$core.bool>(2, 'isExtension', $pb.PbFieldType.QB);
 
@@ -1320,26 +1642,35 @@ class UninterpretedOption_NamePart extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static UninterpretedOption_NamePart _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get namePart => $_getS(0, '');
+  @$pb.TagNumber(1)
   set namePart($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasNamePart() => $_has(0);
+  @$pb.TagNumber(1)
   void clearNamePart() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.bool get isExtension => $_get(1, false);
+  @$pb.TagNumber(2)
   set isExtension($core.bool v) {
     $_setBool(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasIsExtension() => $_has(1);
+  @$pb.TagNumber(2)
   void clearIsExtension() => clearField(2);
 }
 
 class UninterpretedOption extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('UninterpretedOption',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..pc<UninterpretedOption_NamePart>(2, 'name', $pb.PbFieldType.PM,
         subBuilder: UninterpretedOption_NamePart.create)
     ..aOS(3, 'identifierValue')
@@ -1371,60 +1702,86 @@ class UninterpretedOption extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static UninterpretedOption _defaultInstance;
 
+  @$pb.TagNumber(2)
   $core.List<UninterpretedOption_NamePart> get name => $_getList(0);
 
+  @$pb.TagNumber(3)
   $core.String get identifierValue => $_getS(1, '');
+  @$pb.TagNumber(3)
   set identifierValue($core.String v) {
     $_setString(1, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasIdentifierValue() => $_has(1);
+  @$pb.TagNumber(3)
   void clearIdentifierValue() => clearField(3);
 
+  @$pb.TagNumber(4)
   Int64 get positiveIntValue => $_getI64(2);
+  @$pb.TagNumber(4)
   set positiveIntValue(Int64 v) {
     $_setInt64(2, v);
   }
 
+  @$pb.TagNumber(4)
   $core.bool hasPositiveIntValue() => $_has(2);
+  @$pb.TagNumber(4)
   void clearPositiveIntValue() => clearField(4);
 
+  @$pb.TagNumber(5)
   Int64 get negativeIntValue => $_getI64(3);
+  @$pb.TagNumber(5)
   set negativeIntValue(Int64 v) {
     $_setInt64(3, v);
   }
 
+  @$pb.TagNumber(5)
   $core.bool hasNegativeIntValue() => $_has(3);
+  @$pb.TagNumber(5)
   void clearNegativeIntValue() => clearField(5);
 
+  @$pb.TagNumber(6)
   $core.double get doubleValue => $_getN(4);
+  @$pb.TagNumber(6)
   set doubleValue($core.double v) {
     $_setDouble(4, v);
   }
 
+  @$pb.TagNumber(6)
   $core.bool hasDoubleValue() => $_has(4);
+  @$pb.TagNumber(6)
   void clearDoubleValue() => clearField(6);
 
+  @$pb.TagNumber(7)
   $core.List<$core.int> get stringValue => $_getN(5);
+  @$pb.TagNumber(7)
   set stringValue($core.List<$core.int> v) {
     $_setBytes(5, v);
   }
 
+  @$pb.TagNumber(7)
   $core.bool hasStringValue() => $_has(5);
+  @$pb.TagNumber(7)
   void clearStringValue() => clearField(7);
 
+  @$pb.TagNumber(8)
   $core.String get aggregateValue => $_getS(6, '');
+  @$pb.TagNumber(8)
   set aggregateValue($core.String v) {
     $_setString(6, v);
   }
 
+  @$pb.TagNumber(8)
   $core.bool hasAggregateValue() => $_has(6);
+  @$pb.TagNumber(8)
   void clearAggregateValue() => clearField(8);
 }
 
 class SourceCodeInfo_Location extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('SourceCodeInfo.Location',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..p<$core.int>(1, 'path', $pb.PbFieldType.K3)
     ..p<$core.int>(2, 'span', $pb.PbFieldType.K3)
     ..aOS(3, 'leadingComments')
@@ -1455,32 +1812,44 @@ class SourceCodeInfo_Location extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static SourceCodeInfo_Location _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.List<$core.int> get path => $_getList(0);
 
+  @$pb.TagNumber(2)
   $core.List<$core.int> get span => $_getList(1);
 
+  @$pb.TagNumber(3)
   $core.String get leadingComments => $_getS(2, '');
+  @$pb.TagNumber(3)
   set leadingComments($core.String v) {
     $_setString(2, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasLeadingComments() => $_has(2);
+  @$pb.TagNumber(3)
   void clearLeadingComments() => clearField(3);
 
+  @$pb.TagNumber(4)
   $core.String get trailingComments => $_getS(3, '');
+  @$pb.TagNumber(4)
   set trailingComments($core.String v) {
     $_setString(3, v);
   }
 
+  @$pb.TagNumber(4)
   $core.bool hasTrailingComments() => $_has(3);
+  @$pb.TagNumber(4)
   void clearTrailingComments() => clearField(4);
 
+  @$pb.TagNumber(6)
   $core.List<$core.String> get leadingDetachedComments => $_getList(4);
 }
 
 class SourceCodeInfo extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('SourceCodeInfo',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..pc<SourceCodeInfo_Location>(1, 'location', $pb.PbFieldType.PM,
         subBuilder: SourceCodeInfo_Location.create)
     ..hasRequiredFields = false;
@@ -1505,13 +1874,15 @@ class SourceCodeInfo extends $pb.GeneratedMessage {
   static SourceCodeInfo getDefault() => _defaultInstance ??= create()..freeze();
   static SourceCodeInfo _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.List<SourceCodeInfo_Location> get location => $_getList(0);
 }
 
 class GeneratedCodeInfo_Annotation extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo(
       'GeneratedCodeInfo.Annotation',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..p<$core.int>(1, 'path', $pb.PbFieldType.K3)
     ..aOS(2, 'sourceFile')
     ..a<$core.int>(3, 'begin', $pb.PbFieldType.O3)
@@ -1543,36 +1914,50 @@ class GeneratedCodeInfo_Annotation extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static GeneratedCodeInfo_Annotation _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.List<$core.int> get path => $_getList(0);
 
+  @$pb.TagNumber(2)
   $core.String get sourceFile => $_getS(1, '');
+  @$pb.TagNumber(2)
   set sourceFile($core.String v) {
     $_setString(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasSourceFile() => $_has(1);
+  @$pb.TagNumber(2)
   void clearSourceFile() => clearField(2);
 
+  @$pb.TagNumber(3)
   $core.int get begin => $_get(2, 0);
+  @$pb.TagNumber(3)
   set begin($core.int v) {
     $_setSignedInt32(2, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasBegin() => $_has(2);
+  @$pb.TagNumber(3)
   void clearBegin() => clearField(3);
 
+  @$pb.TagNumber(4)
   $core.int get end => $_get(3, 0);
+  @$pb.TagNumber(4)
   set end($core.int v) {
     $_setSignedInt32(3, v);
   }
 
+  @$pb.TagNumber(4)
   $core.bool hasEnd() => $_has(3);
+  @$pb.TagNumber(4)
   void clearEnd() => clearField(4);
 }
 
 class GeneratedCodeInfo extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('GeneratedCodeInfo',
-      package: const $pb.PackageName('google.protobuf'))
+      package: const $pb.PackageName('google.protobuf'),
+      createEmptyInstance: create)
     ..pc<GeneratedCodeInfo_Annotation>(1, 'annotation', $pb.PbFieldType.PM,
         subBuilder: GeneratedCodeInfo_Annotation.create)
     ..hasRequiredFields = false;
@@ -1598,5 +1983,6 @@ class GeneratedCodeInfo extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static GeneratedCodeInfo _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.List<GeneratedCodeInfo_Annotation> get annotation => $_getList(0);
 }