+++ protoc_plugin/lib/src/plugin.pb.dart
@@ -13,7 +13,8 @@ import 'descriptor.pb.dart' as $0;
 
 class Version extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('Version',
-      package: const $pb.PackageName('google.protobuf.compiler'))
+      package: const $pb.PackageName('google.protobuf.compiler'),
+      createEmptyInstance: create)
     ..a<$core.int>(1, 'major', $pb.PbFieldType.O3)
     ..a<$core.int>(2, 'minor', $pb.PbFieldType.O3)
     ..a<$core.int>(3, 'patch', $pb.PbFieldType.O3)
@@ -39,42 +40,59 @@ class Version extends $pb.GeneratedMessage {
   static Version getDefault() => _defaultInstance ??= create()..freeze();
   static Version _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.int get major => $_get(0, 0);
+  @$pb.TagNumber(1)
   set major($core.int v) {
     $_setSignedInt32(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasMajor() => $_has(0);
+  @$pb.TagNumber(1)
   void clearMajor() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.int get minor => $_get(1, 0);
+  @$pb.TagNumber(2)
   set minor($core.int v) {
     $_setSignedInt32(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasMinor() => $_has(1);
+  @$pb.TagNumber(2)
   void clearMinor() => clearField(2);
 
+  @$pb.TagNumber(3)
   $core.int get patch => $_get(2, 0);
+  @$pb.TagNumber(3)
   set patch($core.int v) {
     $_setSignedInt32(2, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasPatch() => $_has(2);
+  @$pb.TagNumber(3)
   void clearPatch() => clearField(3);
 
+  @$pb.TagNumber(4)
   $core.String get suffix => $_getS(3, '');
+  @$pb.TagNumber(4)
   set suffix($core.String v) {
     $_setString(3, v);
   }
 
+  @$pb.TagNumber(4)
   $core.bool hasSuffix() => $_has(3);
+  @$pb.TagNumber(4)
   void clearSuffix() => clearField(4);
 }
 
 class CodeGeneratorRequest extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('CodeGeneratorRequest',
-      package: const $pb.PackageName('google.protobuf.compiler'))
+      package: const $pb.PackageName('google.protobuf.compiler'),
+      createEmptyInstance: create)
     ..pPS(1, 'fileToGenerate')
     ..aOS(2, 'parameter')
     ..a<Version>(3, 'compilerVersion', $pb.PbFieldType.OM,
@@ -104,31 +122,42 @@ class CodeGeneratorRequest extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static CodeGeneratorRequest _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.List<$core.String> get fileToGenerate => $_getList(0);
 
+  @$pb.TagNumber(2)
   $core.String get parameter => $_getS(1, '');
+  @$pb.TagNumber(2)
   set parameter($core.String v) {
     $_setString(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasParameter() => $_has(1);
+  @$pb.TagNumber(2)
   void clearParameter() => clearField(2);
 
+  @$pb.TagNumber(3)
   Version get compilerVersion => $_getN(2);
+  @$pb.TagNumber(3)
   set compilerVersion(Version v) {
     setField(3, v);
   }
 
+  @$pb.TagNumber(3)
   $core.bool hasCompilerVersion() => $_has(2);
+  @$pb.TagNumber(3)
   void clearCompilerVersion() => clearField(3);
 
+  @$pb.TagNumber(15)
   $core.List<$0.FileDescriptorProto> get protoFile => $_getList(3);
 }
 
 class CodeGeneratorResponse_File extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo(
       'CodeGeneratorResponse.File',
-      package: const $pb.PackageName('google.protobuf.compiler'))
+      package: const $pb.PackageName('google.protobuf.compiler'),
+      createEmptyInstance: create)
     ..aOS(1, 'name')
     ..aOS(2, 'insertionPoint')
     ..aOS(15, 'content')
@@ -158,34 +187,47 @@ class CodeGeneratorResponse_File extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static CodeGeneratorResponse_File _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get name => $_getS(0, '');
+  @$pb.TagNumber(1)
   set name($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasName() => $_has(0);
+  @$pb.TagNumber(1)
   void clearName() => clearField(1);
 
+  @$pb.TagNumber(2)
   $core.String get insertionPoint => $_getS(1, '');
+  @$pb.TagNumber(2)
   set insertionPoint($core.String v) {
     $_setString(1, v);
   }
 
+  @$pb.TagNumber(2)
   $core.bool hasInsertionPoint() => $_has(1);
+  @$pb.TagNumber(2)
   void clearInsertionPoint() => clearField(2);
 
+  @$pb.TagNumber(15)
   $core.String get content => $_getS(2, '');
+  @$pb.TagNumber(15)
   set content($core.String v) {
     $_setString(2, v);
   }
 
+  @$pb.TagNumber(15)
   $core.bool hasContent() => $_has(2);
+  @$pb.TagNumber(15)
   void clearContent() => clearField(15);
 }
 
 class CodeGeneratorResponse extends $pb.GeneratedMessage {
   static final $pb.BuilderInfo _i = $pb.BuilderInfo('CodeGeneratorResponse',
-      package: const $pb.PackageName('google.protobuf.compiler'))
+      package: const $pb.PackageName('google.protobuf.compiler'),
+      createEmptyInstance: create)
     ..aOS(1, 'error')
     ..pc<CodeGeneratorResponse_File>(15, 'file', $pb.PbFieldType.PM,
         subBuilder: CodeGeneratorResponse_File.create)
@@ -214,13 +256,18 @@ class CodeGeneratorResponse extends $pb.GeneratedMessage {
       _defaultInstance ??= create()..freeze();
   static CodeGeneratorResponse _defaultInstance;
 
+  @$pb.TagNumber(1)
   $core.String get error => $_getS(0, '');
+  @$pb.TagNumber(1)
   set error($core.String v) {
     $_setString(0, v);
   }
 
+  @$pb.TagNumber(1)
   $core.bool hasError() => $_has(0);
+  @$pb.TagNumber(1)
   void clearError() => clearField(1);
 
+  @$pb.TagNumber(15)
   $core.List<CodeGeneratorResponse_File> get file => $_getList(1);
 }